2002-10-14  Mike Kestner <mkestner@speakeasy.net>
	* glib/Object.cs (GetObject): remove (int) cast

svn path=/trunk/gtk-sharp/; revision=8277

+++ glib/Object.cs
@@ -104,7 +104,7 @@ namespace GLib {
 
 		public static Object GetObject(IntPtr o)
 		{
-			Object obj = (Object)Objects[(int)o];
+			Object obj = (Object)Objects[o];
 			if (obj != null) return obj;
 			return GtkSharp.ObjectManager.CreateObject(o); 
 		}
